Data Structure V2 part 1
I thank thee, Bruninfante, for sharing thy knowledge and enlightening me so I don't have to figure out things thou hast already figured out.

Not everything works as intended yet and not everything is properly updated yet, but that was kind of implied with the "part 1". Aslo documentation is still incomplete or missing in places.